Updated mcce4_tools/mcce4/pdb_conversions.py. Function parse_cif_line is parsing data from earlier columns (as that data is the ideal data) as opposed to data from later columns.

+++ mcce4_tools/mcce4/pdb_conversions.py
@@ -32,20 +32,26 @@ def validate_cif_convert_args(args: Namespace) -> Tuple[Path, Path]:
     if cif_fp.suffix != ".cif":
         sys.exit("Wrong file: Expecting a .cif extension.")
 
-    if not args.pdb_file:
-        pdb_fp = cif_fp.with_suffix(".pdb")
-    else:
-        if args.pdb_file[-4:] != ".pdb":
-            print("Adding '.pdb' extension to output file name.")
-            pdb_fp = Path(args.pdb_file).with_suffix(".pdb")
-        else:
-            pdb_fp = Path(args.pdb_file)
+    if not cif_fp.is_file():
+        sys.exit(f"{cif_fp!s} is not a file.")
 
-    if pdb_fp.exists():
-        if args.overwrite:
-            pdb_fp.unlink()
-        else:
-            sys.exit("Exiting: Output file already exists & overwrite is False.")
+    if not cif_fp.stat().st_size:
+        sys.exit(f"File is empty: {cif_fp!s}.")
+
+    if args.pdb_file:
+        pdb_fp = Path(args.pdb_file)
+        
+        if pdb_fp.suffix != ".pdb":
+            sys.exit("Wrong file: Expecting a .pdb extension.")
+
+        if pdb_fp.exists():
+            if args.overwrite:
+                pdb_fp.unlink()
+            else:
+                sys.exit("Output file already exists.")
+    else:
+        pdb_fp = cif_fp.with_suffix(".pdb")
+        print(f"No output file provided. Using {pdb_fp!s}.")
 
     return cif_fp, pdb_fp
 
@@ -84,7 +90,7 @@ def parse_cif_line(cif_line: str) -> List:
     79 - 80       LString(2)    charge       Charge  on the atom.
     """
     # ATOM   1020 O OXT A LEU A 1 129 ? -10.591 21.716 7.719  0.57 52.55  ? 129  LEU A OXT 1
-    rec, serial, elem, atm, alt, res, _, _, _, _, x, y, z, occ, beta, _, seqnum, _, chn, *_ = cif_line.split()
+    rec, serial, elem, atm, alt, res, chn, _, seqnum, _, x, y, z, occ, beta, *_ = cif_line.split()
     # return fields in pdb format order & floats :
     return [rec, serial, atm, alt, res, chn, seqnum,
             float(x), float(y), float(z), float(occ), float(beta), elem]
@@ -101,6 +107,18 @@ def get_pdb_line(cif_line: str) -> str:
     if not alt or alt == ".":
         alt = " "
 
+    if len(chn) > 1:
+        sys.exit("Error: CIF file chain ID has more than 1 character. Exiting... ")
+    
+    if int(serial) > 99999:
+        sys.exit("Error: CIF file has more than 99999 ATOM coordinates. Exiting... ")
+    
+    if float(beta) > 999.99:
+        sys.exit("Error: CIF file has B-factors > 999.99. Exiting... ")
+
+    if len(elem) > 4:
+        sys.exit("Error: CIF file has residue names that are more than 4 characters long. Exiting... ")
+
     align = "^" if len(atm) < 3 else ">"
     pdbline = (
         f"{rec:<6s}{serial:>5s} {atm:{align}4s}{alt}{res:<3s} {chn}{seqnum:>4s}{' ':4s}"
@@ -111,20 +129,23 @@ def get_pdb_line(cif_line: str) -> str:
 
 def cif_conversion(args: Namespace):
     
-    MAX_PDB_LINES = 99_999
+    MAX_PDB_LINES = 135_000
 
     cif_fp, pdb_fp = validate_cif_convert_args(args)
+    
     # extract the coordinates lines:
-    pattern = re.compile(r"^[ATOM  |HETATM].*$", re.MULTILINE)
-    cif_lines = pattern.findall(Path(cif_fp).read_text())
-    if not cif_lines:
-        sys.exit(f"No coordinate lines found in {cif_fp!s}.")
-
+    with open(cif_fp) as cif:
+        cif_lines = cif.readlines()
+    
+    # Check if the cif file is too big for conversion
     if len(cif_lines) > MAX_PDB_LINES:
-        sys.exit(f"This cif file {cif_fp!s} is to large for conversion to pdb format.")
+        sys.exit("The cif file is to large for conversion to pdb format.")
 
     with open(pdb_fp, "w") as pdb:
         for line in cif_lines:
+            if not line.startswith(("ATOM","HETATM")):
+                continue
+
             pdbline = get_pdb_line(line)
             pdb.write(pdbline)
 
@@ -149,10 +170,10 @@ def cif_to_pdb_parser() -> ArgumentParser:
         help="Input cif file",
     )
     p.add_argument(
-        "--pdb_file",
+        "pdb_file",
         type=str,
-        default="",
-        help="Output pdb file if different from the input file name; Default %(default)s.",
+        nargs="?",
+        help="Optional output pdb file if different from the input file name; Default %(default)s.",
     )
     p.add_argument(
         "--overwrite",
@@ -169,3 +190,7 @@ def to_cif_cli(argv=None):
     cif_conversion(args)
 
     return
+
+
+if __name__ == "__main__":
+    sys.exit(to_cif_cli())
